preparation for update of edit system
