+++ GD-77_new_firmware/firmware/source/functions/fw_settings.c
@@ -24,37 +24,58 @@
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _FW_EDIT_H_
-#define _FW_EDIT_H_
-
-#include "fw_common.h"
-#include "fw_keyboard.h"
-#include "fw_display.h"
-
-#include "UC1701.h"
-
-#include "fw_main.h"
-#include "fw_sound.h"
 #include "fw_settings.h"
 
-#define MODE_ANALOG 1
-#define MODE_DIGITAL 2
+int current_mode = 0;
+int current_frequency[FREQ_COUNT];
+int current_frequency_idx;
 
-#define VHF_MIN 1440000
-#define VHF_MAX 1479999
-#define UHF_MIN 4300000
-#define UHF_MAX 4499999
+void save_value(int idx, int value)
+{
+	int tmp_value=load_value(idx);
+	if (tmp_value!=value)
+	{
+		EEPROM_Write(STORAGE_BASE_ADDRESS+idx*sizeof(int), (uint8_t*)&value, sizeof(int));
+	    vTaskDelay(portTICK_PERIOD_MS * 5);
+	}
+}
 
-void reset_freq_enter_digits();
-int read_freq_enter_digits();
-bool check_frequency(int tmp_frequency);
-bool check_frequency_is_VHF();
-bool check_frequency_is_UHF();
-void update_frequency(int tmp_frequency);
-void update_flags();
-void update_qsodata();
-void update_screen();
-void init_edit();
-void tick_edit();
+int load_value(int idx)
+{
+	int tmp_value=0;
+	EEPROM_Read(STORAGE_BASE_ADDRESS+idx*sizeof(int), (uint8_t*)&tmp_value, sizeof(int));
+	return tmp_value;
+}
 
-#endif /* _FW_EDIT_H_ */
+void save_settings()
+{
+	save_value(0, STORAGE_MAGIC_NUMBER);
+	save_value(1, current_mode);
+	save_value(2, current_frequency_idx);
+	for (int i=0;i<FREQ_COUNT;i++)
+	{
+		save_value(3+i, current_frequency[i]);
+	}
+}
+
+void load_settings()
+{
+	if (load_value(0)==STORAGE_MAGIC_NUMBER)
+	{
+		current_mode = load_value(1);
+		current_frequency_idx = load_value(2);
+		for (int i=0;i<FREQ_COUNT;i++)
+		{
+			current_frequency[i] = load_value(3+i);
+		}
+	}
+	else
+	{
+		current_mode = MODE_DIGITAL;
+		for (int i=0;i<FREQ_COUNT;i++)
+		{
+			current_frequency[i] = VHF_MIN;
+		}
+		current_frequency_idx=0;
+	}
+}
